TDD Harness
I wrote these two paragraphs of sql to make it possible to do basic TDD in Transact-Sql without a utility. As you can see, thes are testing for two schemas that form part of a design. Why do it in Sql? Perhaps you haven't chosen a tool that does what you want, or you have chosen one and your employer won't allow you to use it.

The code is deliberately simple so that you can copy and paste it as you evolve the design. One possible advantage over a utility is that you could convert the code into a stored procedure and run it with Sql Agent, providing an automatic validation that the system still has everything it needs to run.