spObjectExists
This Transact-Sql file creates a new Sql Server stored procedure to test for the existence of objects in the database. It was created as a simple "Test First" utility, to check the database catalog for the items that are to be developed. It is used as follows:

EXEC spObjectExists 'SCHEMA', 'dbo'
EXEC spObjectExists 'VIEW', 'sys', 'all_objects'

...and so on. If the object is found an informational error (Level 10) is raised which won't register as a failure, and if it isn't found the level is 16, so it's easy to tell whether everything is correct.