Create CREATE PROCEDURE [dbo].[TestObjectExists].sql
Procedures will now be prefixed with "Test" so that they sort together in Object Explorer and Intellisense. This version will now find objects in other databases on the server, meaning that you could have a utility database and query from there. This would mean specifying all the arguments though; the parameters have been re-arranged to make it quicker to use. The object type must be specified and so must the object type but if the third argument ("@i_ParentSchemaName") is left blank the schema will default to 'dbo', and if the fourth argument ("@i_DbName") is blank the search defaults to the local database.